+++ crates/zng-app/src/event/events.rs
@@ -1,4 +1,4 @@
-use hashbrown::HashSet;
+use std::collections::HashSet;
 use zng_app_context::app_local;
 use zng_time::INSTANT_APP;
 use zng_txt::Txt;