+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Upgrade all crates to Rust 2024 edition.
 * Many return impl Trait changed lifetimes slightly, some replaced with doc hidden types. In practice all code should still just compile,
   but these are breaking changes.
+* Some public `hashbrown::{HashMap, HashSet}` API replaced with the `std::collections` equivalent.
 
 # 0.13.12